refactor(hooks): clean up imports
